local-power-press: verify if power on really worked
If given board has input GPIO connected to power LED, verify if power on
really worked - and press the button again if it didn't. In some cases,
when power was off for a short duration, the system will remember last
state instead of starting in configured state. Specifically, it will
automatically power on, and pressing the button will actually turn it
off. This fixes the issue.

Do not check the LED before pressing, as it will not always work if the
system is suspended, not off - as the LED blinks then, so it may looks
like on, when in fact it isn't.