Add simple, opportunistic host reservation
Add claim/release commands to testbed-control that helps coordinating
usage between different runners (gitlab, openqa, manual). When reserved
by one, stop the other services to prevent them scheduling any work
(nicer than failing them at the beginning).
For openqa-worker, it's done via isotovideo wrapper, which is started at
the very beginning of the job. But for gitlab-runner, there is no such early
hook, so use pre_get_sources_script (and post_build_script) for this.
This unfortunately leaves few seconds window when gitlab-runner already
handles job, but openqa-worker isn't stopped yet. If openqa job gets
scheduled in this window, it will stop gitlab-runner, failing the job.
Minimize the effect by stopping the gitlab-runner with SIGQUIT - which
should exit gracefully (wait until current job completes).

If that turns out to be an issue in practice, alternative could be
blocking "claim" command (and sufficiently high timeouts in both openqa
and gitlab). But for now leave it as is and see how it works.