Add ipxe support for legacy boot
grub2 is not great at network booting in legacy mode (it lacks many network
drivers and relies on firmware for that). Use iPXE for that. In fact,
iPXE was already used on some hosts on UEFI to chainload grub2-efi (when
firmware/network card lacks network boot support), but use it directly
for legacy boot.
The implementation is very similar to the grub2-efi one:
- have a main config, that loads "settings" config based on the mac
  address
- that "setting" config sets a bunch of variables, including "testbedid"
- load testbed{testbedid}-default.ipxe config that selects default boot
  entry - this file gets symlinked to the currently active boot mode

This requires ipxe-openqa.ipxe file in the testXX dir for openqa -
similar to grub-openqa.cfg. And for custom mode, one needs to create
ipxe.ipxe in the testXX dir.