thor: take a lock when claiming a host
That operation can take a while, especially when waiting for gitlab job
to complete, which sometimes results in a race condition. Specifically:
1. Gitlab job starts
2. OpenQA job starts
3. OpenQA tries to claim the host
4. Claiming first stops gitlab-runner, which waits for the current job
   to complete
5. Gitlab tries to claim the host
6. Claiming tries to stop openqa worker, and this works.

The end result is that gitlab job completes, but the gitlab-runner is
left stopped. If the gitlab job failed, it doesn't call release, so
openqa worker remains stopped too and no further tests will run.

Some of the above could be mitigated by reordering calls in the claim
function, but it only (significantly) narrows the race window, but
doesn't fully eliminate it.

Fix the issue by taking (per-host) lock during claim operation. Release
doesn't need such lock, as release can be done only by the entity that
claimed it earlier, so there is nobody to race against.