Update power-press for new libgpiod version, add pwmchip support
libgpiod3 replaced libgpiod2 and its API is substantially different.
This broke the python bindings after update too. Update to the new API.

But also, allow using gpiochip sysfs API for boards having support for
that. The actual gpiochip needs to be enabled via device tree overlay
(check board manual to see which one is connected to which GPIO), and
then in the hw-control.conf use "gpiochip1" in place of just pin number.