Skip wiping the disk for reinstall, if requested snapshot is still there
If CONFIRM_SNAPSHOT_OR_REINSTALL=1 is set, inspect HDD_1 and check if
that snapshot is still on the disk, instead of restoring it. If it's
there (and there is more than 15% free space in the root pool), skip
restore and set SKIP_INSTALL=1 variable. If it's not there, wipe the
disk as usual (as if HDD_1 was not set).